+++ PerformanceWork.md
@@ -1,0 +1,3 @@
+Get the web socket and rest working in a interface sort of way
+Get Startup and Shutdown of the implementations working as client ID changes
+Get graceful shutdowns
